requirements.txt
juntagrico==1.4.0
git+https://github.com/juntagrico/juntagrico-billing.git@486370894005a8995357e1fec8669cefcac51656

gunicorn==20.0.4
# psycopg2==2.8.4
psycopg2-binary # msch
#django==3.1.5 # TODO full versions here, upgrade?
django-mail-queue==3.2.2
django-oauth-toolkit==1.3.2
pycryptodomex==3.9.7	# fix version of this oidc-provider dependency
django-oidc-provider==0.7.0
whitenoise==5.2.0
dj-database-url==0.5.0
django-storages>=1.1.8
django-admin-sortable2>=0.7.8
boto3>=1.16.0
reportlab==3.6.6  # newer versions seem to lack getStringIO?
